backend/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
celery==5.2.7
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-memcached>=1.3
pytz==2021.3
sqlparse==0.4.2